api: modularize handlers (#84)
* api: modularize handlers

* api: guard nil custom_json (#85)
