+++ cmd/api/watchers/watchers.go
@@ -1,0 +1,14 @@
+package watchers
+
+import (
+	"net/http"
+
+	"github.com/gin-gonic/gin"
+	app "github.com/mark3748/helpdesk-go/cmd/api/app"
+)
+
+func List(a *app.App) gin.HandlerFunc { return func(c *gin.Context) { c.JSON(http.StatusOK, []any{}) } }
+func Add(a *app.App) gin.HandlerFunc  { return func(c *gin.Context) { c.Status(http.StatusCreated) } }
+func Remove(a *app.App) gin.HandlerFunc {
+	return func(c *gin.Context) { c.JSON(http.StatusOK, gin.H{"ok": true}) }
+}
